examples/REST/requirements.txt
requests>=2.31.0
pandas>=2.0.0
rich>=13.0.0
Jinja2>=3.1.0
matplotlib>=3.7.0
fpdf2>=2.7.4
seaborn>=0.12.2
python-dotenv>=0.21.0